.docker/chat_notifs/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
google-crc32c
google-cloud-secret-manager